pyproject.toml
[project]
name = "agent-sandbox"
version = "0.1.0"
description = "CLI tool for creating sandboxed development environments using git worktrees and Docker"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
agent-sandbox = "agent_sandbox.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_sandbox"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests requiring Docker",
]
